+++ im_flutter_sdk_ios/ios/im_flutter_sdk_ios/Sources/im_flutter_sdk_ios/include/im_flutter_sdk_ios/MethodKeys.h
@@ -115,6 +115,8 @@ static NSString *const ChatRemoveMessagesFromServerWithTs = @"removeMessagesFrom
 
 static NSString *const GetConversationsFromServerWithCursor = @"getConversationsFromServerWithCursor";
 static NSString *const GetPinnedConversationsFromServerWithCursor = @"getPinnedConversationsFromServerWithCursor";
+// 4.25.0
+static NSString *const GetConversationsFromDBWithCursor = @"getConversationsFromDBWithCursor";
 static NSString *const PinConversation = @"pinConversation";
 static NSString *const modifyMessage = @"modifyMessage";
 static NSString *const downloadAndParseCombineMessage = @"downloadAndParseCombineMessage";